+++ services/frontend/src/app/components/QuestionOfTheDayCard.tsx
