+++ services/frontend/src/app/components/WeeklySummaryCard.tsx
