--- services/frontend/src/app/home/page.tsx
@@ -1,24 +1,0 @@
-"use client";
-
-import { HistoryTable } from "./components/HistoryTable";
-import { QuestionOfTheDayCard } from "./components/QuestionOfTheDayCard";
-import { WeeklySummaryCard } from "./components/WeeklySummaryCard";
-import { historiesStub } from "./stub/historyStub";
-import { questionStub } from "./stub/questionStub";
-
-const page = () => {
-  return (
-    <div>
-      <h1 className="my-4 text-3xl font-semibold">Welcome Back, Wei Jun!</h1>
-
-      <div className="flex w-full gap-x-5">
-        <WeeklySummaryCard />
-        <QuestionOfTheDayCard question={questionStub} />
-      </div>
-      <h1 className="my-4 text-3xl font-semibold">Past Interview</h1>
-      <HistoryTable histories={historiesStub} />
-    </div>
-  );
-};
-
-export default page;
