+++ services/frontend/src/app/stub/questionStub.ts
